Refactor Models
