+++ configs/model/classification/pointnet++.yaml
@@ -2,7 +2,7 @@ defaults:
   - optimizer: adam
   - lr_scheduler: ReduceLROnPlateau
 
-_target_: src.models.openpoints_module.OpenPointsModule
+_target_: src.models.classification.openpoints_module.OpenPointsModule
 
 optimizer:
   lr: 0.001
